renaming and using Models
