+++ php/libraries/Data/Models/TransferObjects/Candidate.php
@@ -7,9 +7,9 @@
  * 
  * See Candidate_DAO for persistance.
  */
-namespace LORIS\Data\DTO;
+namespace LORIS\Data\Models\TransferObjects;
 
-class Candidate extends \LORIS\Data\DTO
+class Candidate extends \LORIS\Data\Models\TransferObject
 {
     /**
      * This data is directly mapped to the columns of database table.
